requirements.txt
discord.py
python-dotenv
aiosqlite
watchdog
